integer-based min-distance for NMR measurements; float parsing fix
Former-commit-id: 8783105b67607bca3d36b5e4838358a8a74ec854

+++ Jmol/src/org/jmol/api/JmolNMRInterface.java
@@ -82,7 +82,7 @@ public interface JmolNMRInterface {
 
   public JmolList<Object> getTensorInfo(String tensorType, String infoType, BS bs);
 
-  public Map<String, Float> getMinDistances(MeasurementData md);
+  public Map<String, Integer> getMinDistances(MeasurementData md);
 
   public boolean getState(SB sb);
